requirements.txt
# PDF Image Extraction Requirements
PyMuPDF==1.23.14
Pillow==10.1.0
pdf2image==1.16.3
opencv-python==4.8.1.78

# Note: You'll also need poppler-utils installed on your system
# macOS: brew install poppler
# Ubuntu/Debian: sudo apt-get install poppler-utils
